Ny tjeneste for å utlede manglende søkte perioder for behandlinger som ikke krever sammenhengende uttak/fri utsettelse
