+++ src/main/java/no/nav/foreldrepenger/regler/uttak/fastsetteperiode/grunnlag/Søknadstype.java
@@ -8,4 +8,8 @@ public enum Søknadstype {
     public boolean gjelderTerminFødsel() {
         return this == TERMIN || this == FØDSEL;
     }
+
+    public boolean gjelderAdopsjon() {
+        return !gjelderTerminFødsel();
+    }
 }
