+++ src/main/java/no/nav/foreldrepenger/regler/uttak/felles/grunnlag/Periode.java
@@ -32,6 +32,7 @@ public class Periode {
         return overlapper(periode.getFom()) || overlapper(periode.getTom()) || erOmsluttetAv(periode);
     }
 
+    //Også true hvis perioden er lik
     public boolean erOmsluttetAv(LukketPeriode periode) {
         return !periode.getFom().isAfter(fom) && !periode.getTom().isBefore(tom);
     }
